requirements.txt
rembg
pillow
fastapi
onnxruntime
torch
transformers
trimesh 
pyrender 
pyglet 
numpy 
Pillow
pydantic-settings
uvicorn
python-multipart
peft
torchcodec
einops
timm
ftfy
diffusers
faiss-cpu
rembg[gpu]
ftfy
pyngrok
soundfile
librosa
unsloth
qwen_vl_utils
websockets
open-clip-torch
git+https://github.com/huggingface/diffusers.git
huggingface_hub
